refactor to components/FlashCards folder
