+++ packages/frontend/src/components/FlashCards/index.ts
@@ -1,0 +1,1 @@
+export * from "./FlashCards.vue";
